+++ stroom-dashboard-server/src/main/java/stroom/dashboard/server/SearchBusPollActionHandler.java
@@ -81,7 +81,7 @@ class SearchBusPollActionHandler extends AbstractTaskHandler<SearchBusPollAction
             if (LOGGER.isDebugEnabled()) {
                 final StringBuilder sb = new StringBuilder(
                         "Only the following search queries should be active for session '");
-                sb.append(action.getUserToken());
+                sb.append(action.getUserIdentity());
                 sb.append("'\n");
                 for (final DashboardQueryKey queryKey : action.getSearchActionMap().keySet()) {
                     sb.append("\t");
@@ -90,7 +90,7 @@ class SearchBusPollActionHandler extends AbstractTaskHandler<SearchBusPollAction
                 LOGGER.debug(sb.toString());
             }
 
-            final String searchSessionId = action.getUserToken() + "_" + action.getApplicationInstanceId();
+            final String searchSessionId = action.getUserIdentity() + "_" + action.getApplicationInstanceId();
             final ActiveQueries activeQueries = activeQueriesManager.get(searchSessionId);
             final Map<DashboardQueryKey, SearchResponse> searchResultMap = new HashMap<>();