+++ stroom-pipeline/src/main/java/stroom/pipeline/server/task/SteppingTask.java
@@ -21,6 +21,7 @@ import stroom.pipeline.shared.StepType;
 import stroom.pipeline.shared.SteppingFilterSettings;
 import stroom.pipeline.shared.SteppingResult;
 import stroom.query.api.v2.DocRef;
+import stroom.security.shared.UserIdentity;
 import stroom.streamstore.shared.FindStreamCriteria;
 import stroom.streamstore.shared.StreamType;
 import stroom.util.task.ServerTask;
@@ -41,8 +42,8 @@ public class SteppingTask extends ServerTask<SteppingResult> {
     private DocRef pipeline;
     private Map<String, String> code;
 
-    public SteppingTask(final String userToken) {
-        super(null, userToken);
+    public SteppingTask(final UserIdentity userIdentity) {
+        super(null, userIdentity);
     }
 
     public FindStreamCriteria getCriteria() {