+++ stroom-security-server/src/main/java/stroom/security/server/SessionListListener.java
@@ -25,12 +25,12 @@ import org.springframework.stereotype.Component;
 import stroom.entity.shared.BaseCriteria;
 import stroom.entity.shared.BaseResultList;
 import stroom.feed.server.UserAgentSessionUtil;
-import stroom.security.UserTokenUtil;
-import stroom.security.shared.UserRef;
+import stroom.security.ProcessingUserIdentity;
+import stroom.security.shared.UserIdentity;
 import stroom.servlet.SessionDetails;
 import stroom.servlet.SessionListService;
+import stroom.task.server.ExtendedFindTaskCriteria;
 import stroom.task.server.TaskManager;
-import stroom.task.shared.FindTaskCriteria;
 import stroom.task.shared.TerminateTaskProgressAction;
 import stroom.util.spring.StroomBeanStore;
 import stroom.util.task.TaskIdFactory;
@@ -86,11 +86,11 @@ public class SessionListListener implements HttpSessionListener, SessionListServ
             // Manually set the id as we are invoking a UI Action Task
             // directly
             final String sessionId = event.getSession().getId();
-            final FindTaskCriteria criteria = new FindTaskCriteria();
+            final ExtendedFindTaskCriteria criteria = new ExtendedFindTaskCriteria();
             criteria.setSessionId(sessionId);
             final TerminateTaskProgressAction action = new TerminateTaskProgressAction(
                     "Terminate session: " + sessionId, criteria, false);
-            action.setUserToken(UserTokenUtil.INTERNAL_PROCESSING_USER_TOKEN);
+            action.setUserIdentity(ProcessingUserIdentity.INSTANCE);
             action.setId(TaskIdFactory.create());
 
             final TaskManager taskManager = getTaskManager();
@@ -108,12 +108,11 @@ public class SessionListListener implements HttpSessionListener, SessionListServ
         for (final HttpSession httpSession : sessionMap.values()) {
             final SessionDetails sessionDetails = new SessionDetails();
 
-            final UserRef user = UserRefSessionUtil.get(httpSession);
-            if (user != null) {
-                sessionDetails.setUserName(user.getName());
+            final UserIdentity userIdentity = UserIdentitySessionUtil.get(httpSession);
+            if (userIdentity != null) {
+                sessionDetails.setUserName(userIdentity.getId());
             }
 
-            sessionDetails.setId(httpSession.getId());
             sessionDetails.setCreateMs(httpSession.getCreationTime());
             sessionDetails.setLastAccessedMs(httpSession.getLastAccessedTime());
             sessionDetails.setLastAccessedAgent(UserAgentSessionUtil.get(httpSession));